requirements.txt
# Python dependencies for NixOS GUI fallback server
# Used when Node.js is not available

# Web framework
flask>=3.0.0
gunicorn>=21.2.0

# JWT support
PyJWT>=2.8.0

# SSL/TLS
pyOpenSSL>=23.3.0

# WebSocket support
websockets>=12.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
requests>=2.31.0

# Development tools
python-dotenv>=1.0.0
watchdog>=3.0.0

# Security
cryptography>=41.0.0

# Utilities
click>=8.1.0
colorama>=0.4.6